Refactored communication and created color picker
