+++ src/serverCommunication.js
@@ -1,0 +1,15 @@
+import axios from "axios";
+
+const CHRISTMAS_TREE_URL = process.env.REACT_APP_CHRISTMAS_TREE_URL;
+
+function postColorRequest(color) {
+    axios
+        .post(CHRISTMAS_TREE_URL + "/setRgbColor/", {
+            color: color,
+        })
+        .catch(function (error) {
+            console.log(error);
+        });
+}
+
+export { postColorRequest };
